+++ pyleecan/Generator/ClassesRef/Simulation/EEC.csv
@@ -1,9 +1,10 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constante Name,Constante Value,Description classe
 type_skin_effect,-,Skin effect for resistance and inductance,1,int,1,,,,Simulation,,get_machine_from_parent,VERSION,1,Equivalent Electrical Circuit abstract class
-OP,-,Operating Point,,OP,None,,,,,,,,,
+OP,-,Operating Point,,OP,None,,,,,,comp_R1,,,
 Tsta,deg Celsius,Average stator temperature for operational EEC calculation,0,float,20,,,,,,,,,
 Trot,deg Celsius,Average rotor temperature for operational EEC calculation,0,float,20,,,,,,,,,
 Xkr_skinS,,Skin effect coefficient for resistances at stator side,,float,None,,,,,,,,,
 Xke_skinS,,Skin effect coefficient for inductances at stator side,,float,None,,,,,,,,,
 Xkr_skinR,,Skin effect coefficient for resistances at rotor side,,float,None,,,,,,,,,
 Xke_skinR,,Skin effect coefficient for inductances at rotor side,,float,None,,,,,,,,,
+R1,Ohm,Stator phase resistance,,float,None,,,,,,,,,
