+++ pyleecan/Generator/ClassesRef/Simulation/EEC_PMSM.csv
@@ -7,8 +7,7 @@ Phid,Wb,Stator winding flux along d-axis,,float,None,,,,,,comp_torque_sync_rel,,
 Phiq,Wb,Stator winding flux along q-axis,,float,None,,,,,,comp_BEMF_harmonics,,,,
 Phid_mag,Wb,Stator winding flux along d-axis in open-circuit (rotor flux linkage),,float,None,,,,,,comp_Ld,,,,
 Phiq_mag,Wb,Stator winding flux along q-axis in open-circuit (rotor flux linkage),,float,None,,,,,,comp_Lq,,,,
-R1,Ohm,Stator phase resistance,,float,None,,,,,,comp_Phid,,,,
+,,,,,,,,,,,comp_Phid,,,,
 ,,,,,,,,,,,comp_Phiq,,,,
 ,,,,,,,,,,,comp_Phid_mag,,,,
 ,,,,,,,,,,,comp_Phiq_mag,,,,
-,,,,,,,,,,,comp_R1,,,,
