+++ pyleecan/Generator/ClassesRef/Simulation/EEC_SCIM.csv
@@ -5,7 +5,6 @@ Im,Arms,Magnetizing current,,float,None,,,,,,comp_joule_losses,,,,
 If,Arms,Iron loss current,,float,None,,,,,,solve_elementary,,,,
 U1,Vrms,Stator phase voltage,,float,None,,,,,,_comp_flux_mean,,,,
 U2,Vrms,Rotor phase voltage,,float,None,,,,,,_comp_Lm_FEA,,,,
-R1,Ohm,Stator phase resistance,,float,None,,,,,,,,,,
 L1,H,Stator phase inductance,,float,None,,,,,,,,,,
 R2,Ohm,Rotor phase resistance,,float,None,,,,,,,,,,
 L2,H,Rotor phase inductance,,float,None,,,,,,,,,,
